requirements.txt
# Core dependencies
numpy
pandas
pyyaml

# LLM and API
langchain-ollama
requests

# Visualization (for ml_model.py)
matplotlib
seaborn

# Additional dependencies
PyPDF2
python-docx


#python3 -m pip install -r requirements.txt
#python3 -m pip list